refactor the serialization
