+++ src/core/core/rpc_message.cpp
@@ -38,7 +38,7 @@
 # include <dsn/internal/network.h>
 
 # ifdef DSN_ENABLE_THRIFT_RPC
-# include <dsn/idl/thrift_rpc.h>
+# include <dsn/internal/thrift_parser.h>
 # endif
 
 # include "task_engine.h"
